+++ CHANGELOG.md
@@ -15,3 +15,7 @@
 - Minor: 
     - No linter warnings when building with `webpack`
     - chromajs updated to 1.3.4
+
+#v1.2.0 in progress..
+- Optional labels in `colorBar` control (see `example_VectorFieldAnim_ColorBar.html`)
+- Better colorBar styles and more options